* linkers/Linker.nw: LP split stuff, use short instead of char/byte for enum more consistently

+++ linkers/5l/globals.c
@@ -11,7 +11,7 @@ long	HEADR; 		/* length of header */
 /*e: global HEADR */
 /*s: global HEADTYPE */
 // enum<headtype>
-long	HEADTYPE = -1; /* type of header */
+short	HEADTYPE = -1; /* type of header */
 /*e: global HEADTYPE */
 /*s: global INITDAT */
 long	INITDAT = -1; /* data location */
